changed user id from long to int

+++ LoginDemo/demo/src/main/java/comp31/demo/model/User.java
@@ -22,13 +22,13 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name="USER_ID")
-    private Long userId;
+    private Integer userId;
     @Column(name="USER_NAME")
     private String userName;
     @Column(name="USER_ROLE")
     private String userRole;
     
-    public Long getId() {
+    public Integer getId() {
 		return userId;
     }
     public String getName() {
