+++ LoginDemo/demo/src/main/java/comp31/demo/repo/ShowRepo.java
@@ -6,14 +6,13 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 import comp31.demo.model.Show;
-import comp31.demo.model.User;
 
 //Nick Best
 
 @Repository
 public interface ShowRepo extends CrudRepository <Show, Long> {
     public List<Show> findAll();
-    public List<Show> findByUserId(Long userId);
+    public List<Show> findByUserId(Integer userId);
     //Show Repo goes here
     
 }
